backend/requirements.txt
fastapi
requests
urllib3
langchain
langchain_openai
langchain_community
uvicorn
python-dotenv
pydantic_settings
pypdf
supabase
python-multipart
langsmith
langgraph
pymongo
boto3
motor
pypdf
chromadb